requirements.txt
requests
lxml
python-dotenv